Move TutorialPageThree into its own file
